@@ -1,5 +1,5 @@
 import {CommonModule} from '@angular/common';
-import {HttpClientModule} from '@angular/common/http';
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {NgbDropdownModule, NgbNavModule, NgbTooltipModule} from '@ng-bootstrap/ng-bootstrap';
@@ -21,7 +21,6 @@ import {FourPaneEditorComponent} from './four-pane-editor/four-pane-editor.compo
   ],
   imports: [
     CommonModule,
-    HttpClientModule,
     FormsModule,
     NgbNavModule,
     NgbDropdownModule,
@@ -31,6 +30,7 @@ import {FourPaneEditorComponent} from './four-pane-editor/four-pane-editor.compo
     SharedModule,
     ModalModule,
   ],
+  providers: [provideHttpClient(withInterceptorsFromDi())],
 })
 export class EditorModule {
 }
